Split body-highlighter runtime surface into ./bodymap subpath
The main barrel is a deep `export *` chain ending at
Workout/index.ts, which re-exported BodyMap, BodyMapDetailPanel,
TrainingStatusPage, and the muscleTaxonomy value helpers (incl. the
MuscleGroup/SimpleMuscleGroup enums) as runtime values. BodyMap.tsx
and muscleTaxonomy.ts both import `react-native-body-highlighter` at
module top-level, so importing anything from the barrel (e.g.
`{ Metric }`) statically pulled that native SVG package into every
consumer's module graph. `sideEffects: false` can't help a static
re-export.

Move those value exports behind a new `@titan-design/react-ui/bodymap`
subpath (src/bodymap.ts) and keep only type-only re-exports on the
barrel (erased at build -> zero runtime pull). TrainingStatusPage's
values move too since it imports BodyMap at runtime.

Proof: after build, `grep -c react-native-body-highlighter dist/index.js`
is 0 (was nonzero); dist/bodymap.js is 1.

Breaking export change -> version 0.2.7 -> 0.3.0. Publish (Phase B) and
the voltras-mcp consumer migration (Phase C) are follow-ups.
