+++ packages/ui/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@titan-design/react-ui",
-  "version": "0.2.7",
+  "version": "0.3.0",
   "description": "Cross-platform design system built on Gluestack UI",
   "author": "Henry Jewkes",
   "license": "MIT",
@@ -24,6 +24,12 @@
       "types": "./dist/index.d.ts",
       "import": "./dist/index.mjs",
       "require": "./dist/index.js"
+    },
+    "./bodymap": {
+      "react-native": "./src/bodymap.ts",
+      "types": "./dist/bodymap.d.ts",
+      "import": "./dist/bodymap.mjs",
+      "require": "./dist/bodymap.js"
     },
     "./theme": {
       "react-native": "./src/theme/index.ts",
