+++ packages/ui/tsup.config.ts
@@ -7,6 +7,7 @@ const __dirname = path.dirname(fileURLToPath(import.meta.url))
 export default defineConfig({
   entry: {
     index: 'src/index.ts',
+    bodymap: 'src/bodymap.ts',
     'theme/index': 'src/theme/index.ts',
     'theme/tokens-css': 'src/theme/tokens-css.ts',
   },
